solve the database issue
